*** empty log message ***
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@97949 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0ca2ec4d82edb4df4e4d8cb72c7ed2b69e7bc4e4

+++ Online/Online/ROLogger/src/MessageLine.cpp
@@ -26,7 +26,7 @@ int MessageLine::analyze() const {
     }
   }
   m_mess = m_buff.find(":",NOD_START);
-  if ( m_mess != std::string::npos ) m_mess += 2;
+  if ( m_mess != (int)std::string::npos ) m_mess += 2;
   return m_type=OTHER;
 }
